First pass at content negotiation
First use a regular expression to extract the q values of the content
types we are interested in. Then find the right q value to go with both
the JSON and form encoded type, going from more precise to less precise.

If both content types are unaccepted return an error, otherwise return
whichever has the highest q value. Ties are decided in favour of JSON.