Strict input validation and reorder of steps
Start by checking if a code was sent with the request, as before. If a
code is supplied by the client, verify it and stop.

Collect all the client supplied data from the query string. Exit if any
data contains an unexpected value. After this we branch:

If a password was submitted the user has already accepted the data.
Check both the password and the CSRF token, then send the user back to
the client with a valid code.

If no password was submitted we need to ask the user for it.