Do not implode booleans into a string
I originally misread what FILTER_REQUIRE_ARRAY did. It does not return
false if a single item in the array does not match, but sets that array
item to false instead.

This adds a check and errors out if a scope is set to false instead of
going ahead and imploding the array.